This version of the stack is able to stablish a full call setup and media setup against itsef. +Moved mux tables from Session to ChannelFactory +Added capability handling to Channelfactory +Added channel opening and mux table transfer negotiations to ControlChannel
git-svn-id: https://asteriskvideo.svn.sourceforge.net/svnroot/asteriskvideo@11 57fa1218-b025-0410-bec8-92b503356f72

+++ libh324m/src/H324CCSRLayer.cpp
@@ -31,9 +31,9 @@
 H324CCSRLayer::H324CCSRLayer() : sdu(255),ccsrl(255)
 {
 	//Initialize variables
-	lastsn = -1;
+	lastsn = (DWORD) -1;
 	sentsn = 0;
-	cmdsn = -1;
+	cmdsn = (DWORD) -1;
 	cmd = NULL;
 	waiting = false;
 	isPDU = false;
